requirements.txt
Django>=4.2
djangorestframework>=3.14
drf-yasg>=1.21
pytest>=7.4
pytest-django>=4.5
python-dateutil>=2.8
pytz>=2023.3
psycopg2-binary>=2.9